Feature/2.3.x kart saklama

+++ admin/model/extension/payment/iyzico_checkout_form.php
@@ -2,10 +2,8 @@
 
 class ModelExtensionPaymentIyzicoCheckoutForm extends Model {
 
-    /**
-     * Create tables
-     */
     public function install() {
+		if(!isset($this->session->data['iyzico_update'])){
         $this->db->query("
 			CREATE TABLE IF NOT EXISTS `" . DB_PREFIX . "iyzico_order` (
 			`iyzico_order_id` INT(11) NOT NULL AUTO_INCREMENT,
@@ -38,31 +36,43 @@ class ModelExtensionPaymentIyzicoCheckoutForm extends Model {
 				ADD COLUMN `card_key` VARCHAR(50),
 				ADD COLUMN `iyzico_api` VARCHAR(100),
 				ENGINE=MyISAM DEFAULT COLLATE=utf8_general_ci;");
-    }
+		}	
+		$file = DIR_APPLICATION . 'iyzico_checkout_form.sql';
+		if (!file_exists($file)) { 
+			}
+		$lines = file($file);
+		if ($lines) {
+			$sql = '';
 
-    /**
-     * Drop tables
-     */
+			foreach($lines as $line) {
+				if ($line && (substr($line, 0, 2) != '--') && (substr($line, 0, 1) != '#')) {
+					$sql .= $line;
+  
+					if (preg_match('/;\s*$/', $line)) {
+						$sql = str_replace("INSERT INTO `oc_", "INSERT INTO `" . DB_PREFIX, $sql);
+						$this->db->query($sql);
+						$sql = '';
+					}
+				}
+			}
+		}
+ }
+	
     public function uninstall() {
+		if(!isset($this->session->data['iyzico_update'])){
         $this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "iyzico_order`;");
         $this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "iyzico_order_refunds`;");
         $this->db->query("ALTER TABLE `" . DB_PREFIX . "customer` DROP COLUMN card_key;");
-		$this->db->query("ALTER TABLE `" . DB_PREFIX . "customer` DROP COLUMN iyzico_api;");
+        $this->db->query("ALTER TABLE `" . DB_PREFIX . "customer` DROP COLUMN iyzico_api;");
+		}
+		$this->db->query("DELETE FROM `" . DB_PREFIX . "modification` WHERE code='iyzico_checkout_form'");	
     }
 
-    /**
-     * Log
-     * 
-     */
     public function logger($message) {
         $log = new Log('iyzico_checkout_form.log');
         $log->write($message);
     }
 
-    /**
-     * Create order entry for iyzico transactions
-     * 
-     */
     public function createOrderEntry($data) {
 
         $query_string = "INSERT INTO " . DB_PREFIX . "iyzico_order SET";
@@ -77,10 +87,6 @@ class ModelExtensionPaymentIyzicoCheckoutForm extends Model {
         return $this->db->getLastId();
     }
 
-    /**
-     * Update order details for iyzico transactions
-     * 
-     */
     public function updateOrderEntry($data, $id) {
 
         $query_string = "UPDATE " . DB_PREFIX . "iyzico_order SET";
@@ -94,74 +100,96 @@ class ModelExtensionPaymentIyzicoCheckoutForm extends Model {
         return $this->db->query($query_string);
     }
 
-    /**
-     * Check Version
-     *
-     */
     public function versionCheck($opencart, $iyzico) {
 
         $ch = curl_init();
         curl_setopt($ch, CURLOPT_URL, 'http://iyzico.kahvedigital.com/version');
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-		curl_setopt($ch, CURLOPT_POSTFIELDS, "opencart=$opencart&iyzico=$iyzico&type=opencart");
+        curl_setopt($ch, CURLOPT_POSTFIELDS, "opencart=$opencart&iyzico=$iyzico&type=opencart");
         $response = curl_exec($ch);
         $response = json_decode($response, true);
         return $response;
     }
 
-    /**
-     * update version
-     */
     public function update($version_updatable) {
+
+        function recurse_copy($src, $dst) {
+            $dir = opendir($src);
+            @mkdir($dst);
+            while (false !== ( $file = readdir($dir))) {
+                if (( $file != '.' ) && ( $file != '..' )) {
+                    if (is_dir($src . '/' . $file)) {
+                        recurse_copy($src . '/' . $file, $dst . '/' . $file);
+                    } else {
+                        copy($src . '/' . $file, $dst . '/' . $file);
+                    }
+                }
+            }
+            closedir($dir);
+        }
+
+        function rrmdir($dir) {
+            if (is_dir($dir)) {
+                $objects = scandir($dir);
+                foreach ($objects as $object) {
+                    if ($object != "." && $object != "..") {
+                        if (filetype($dir . "/" . $object) == "dir")
+                            rrmdir($dir . "/" . $object);
+                        else
+                            unlink($dir . "/" . $object);
+                    }
+                }
+                reset($objects);
+                rmdir($dir);
+            }
+        }
 
         $ch = curl_init();
         curl_setopt($ch, CURLOPT_URL, 'http://iyzico.kahvedigital.com/update');
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-        -
-                curl_setopt($ch, CURLOPT_POSTFIELDS, "new_version=$version_updatable");
+        curl_setopt($ch, CURLOPT_POSTFIELDS, "new_version=$version_updatable");
         $response = curl_exec($ch);
         $response = json_decode($response, true);
-        $arrlength = count($response);
 
-        for ($x = 0; $x < $arrlength; $x++) {
-            $status = $response[$x]['status'];
-            if ($status == 1) {
-                $def_dir = $response[$x]['def_dir'];
+        $serveryol = $_SERVER['DOCUMENT_ROOT'];
+        $ch = curl_init();
+        $source = $response['file_dest'];
+        curl_setopt($ch, CURLOPT_URL, $source);
+        curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
+        $data = curl_exec($ch);
+        curl_close($ch);
 
-                switch ($def_dir) {
-                    case 'catalog':
-                        $def = DIR_CATALOG;
-                        break;
-                    case 'admin':
-                        $def = DIR_APPLICATION;
-                        break;
-                    case 'system':
-                        $def = DIR_SYSTEM;
-                        break;
-					default :
-						 $def ='';
-						 break;
-                }
-                $file_dest = $response[$x]['file_dest'];
-                $new_file = $response[$x]['new_file'];
+        $foldername = $response['version_name'];
+        $fullfoldername = $serveryol . '/' . $foldername;
+		if(!file_exists($fullfoldername)){
+		 mkdir($fullfoldername);
+		}
+        if (file_exists($fullfoldername)) {
+            $unzipfilename = 'iyzicoupdated.zip';
+            $file = fopen($fullfoldername . '/' . $unzipfilename, "w+");
+            fputs($file, $data);
+            fclose($file);
 
-                if ($new_file == 0) {
-                    if (file_exists($def.$file_dest)) {
-                        $fh = fopen($def.$file_dest, 'w');
-                        $content = $response[$x]['content'];
-                        fwrite($fh, $content);
-                        fclose($fh);
-                    }else{
-						return 0;
-					}
-                } else if ($new_file == 1) {
-                    $fh = fopen($def.$file_dest, 'wb');
-                    $content = $response[$x]['content'];
-                    fwrite($fh, $content);
-                    fclose($fh);
-                }
+            $path = pathinfo(realpath($fullfoldername . '/' . $unzipfilename), PATHINFO_DIRNAME);
+            $zip = new ZipArchive;
+            $res = $zip->open($fullfoldername . '/' . $unzipfilename);
+            if ($res === TRUE) {
+                $zip->extractTo($path);
+                $zip->close();
+                $zip_name_folder = $response['zip_name_folder'];
+
+                recurse_copy($fullfoldername . '/' . $zip_name_folder . '/admin', DIR_APPLICATION);
+                recurse_copy($fullfoldername . '/' . $zip_name_folder . '/catalog', DIR_CATALOG);
+                recurse_copy($fullfoldername . '/' . $zip_name_folder . '/system', DIR_SYSTEM);
+
+                rrmdir($fullfoldername);
+            } else {
+                return 0;
             }
+        } else {
+            return 0;
         }
+		$this->load->controller('extension/modification/refresh');
         return 1;
     }
